ridi service 분해 후 재배치
